oc-flash: Remove dead code
Remove dead code which generates warnings with latest gcc.

The function flash_op_verbose() is almost identical to flash_op(),
except for a few extra printf which don't really add much.

The function my_test() is dead code. It's likely a left-over from
development. And it can be found in the git history if needed in the
future.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>